refactor net mods
