+++ open-coroutine-core/src/monitor/mod.rs
@@ -2,8 +2,8 @@ use crate::common::Current;
 use crate::constants::CoroutineState;
 use crate::coroutine::suspender::SimpleSuspender;
 use crate::coroutine::StateCoroutine;
-use crate::event_loop::EventLoops;
 use crate::monitor::node::TaskNode;
+use crate::net::event_loop::EventLoops;
 use crate::scheduler::SchedulableCoroutine;
 use once_cell::sync::{Lazy, OnceCell};
 use open_coroutine_timer::TimerList;
