+++ open-coroutine-core/src/net/mod.rs
@@ -1,0 +1,11 @@
+pub mod config;
+
+#[allow(
+    dead_code,
+    clippy::cast_possible_wrap,
+    clippy::cast_possible_truncation,
+    clippy::cast_sign_loss,
+    trivial_numeric_casts
+)]
+pub mod event_loop;
+mod selector;
