--- open-coroutine-core/src/pool/blocker.rs
@@ -1,6 +1,0 @@
-use std::fmt::Debug;
-use std::time::Duration;
-
-pub trait Blocker: Debug {
-    fn block(&self, time: Duration);
-}
