+++ open-coroutine-hooks/src/unix/mod.rs
@@ -29,7 +29,7 @@ macro_rules! impl_read_hook {
             let error_kind = std::io::Error::last_os_error().kind();
             if error_kind == std::io::ErrorKind::WouldBlock {
                 //wait read event
-                _ = open_coroutine_core::event_loop::EventLoops::wait_read_event(
+                _ = open_coroutine_core::net::event_loop::EventLoops::wait_read_event(
                     socket,
                     Some(std::time::Duration::from_millis(10)),
                 );
@@ -71,7 +71,7 @@ macro_rules! impl_expected_read_hook {
             let error_kind = std::io::Error::last_os_error().kind();
             if error_kind == std::io::ErrorKind::WouldBlock {
                 //wait read event
-                if open_coroutine_core::event_loop::EventLoops::wait_read_event(
+                if open_coroutine_core::net::event_loop::EventLoops::wait_read_event(
                     socket,
                     Some(std::time::Duration::from_millis(10)),
                 )
@@ -145,7 +145,7 @@ macro_rules! impl_expected_batch_read_hook {
             let error_kind = std::io::Error::last_os_error().kind();
             if error_kind == std::io::ErrorKind::WouldBlock {
                 //wait read event
-                if open_coroutine_core::event_loop::EventLoops::wait_read_event(
+                if open_coroutine_core::net::event_loop::EventLoops::wait_read_event(
                     socket,
                     Some(std::time::Duration::from_millis(10)),
                 )
@@ -191,7 +191,7 @@ macro_rules! impl_expected_write_hook {
             let error_kind = std::io::Error::last_os_error().kind();
             if error_kind == std::io::ErrorKind::WouldBlock {
                 //wait write event
-                if open_coroutine_core::event_loop::EventLoops::wait_write_event(
+                if open_coroutine_core::net::event_loop::EventLoops::wait_write_event(
                     socket,
                     Some(std::time::Duration::from_millis(10)),
                 )
@@ -265,7 +265,7 @@ macro_rules! impl_expected_batch_write_hook {
             let error_kind = std::io::Error::last_os_error().kind();
             if error_kind == std::io::ErrorKind::WouldBlock {
                 //wait write event
-                if open_coroutine_core::event_loop::EventLoops::wait_write_event(
+                if open_coroutine_core::net::event_loop::EventLoops::wait_write_event(
                     socket,
                     Some(std::time::Duration::from_millis(10)),
                 )
