+++ open-coroutine-hooks/src/unix/socket.rs
@@ -3,7 +3,7 @@ use once_cell::sync::Lazy;
 use open_coroutine_core::common::Current;
 use open_coroutine_core::constants::Syscall;
 use open_coroutine_core::coroutine::StateCoroutine;
-use open_coroutine_core::event_loop::EventLoops;
+use open_coroutine_core::net::event_loop::EventLoops;
 use std::ffi::c_void;
 use std::io::Error;
 
@@ -16,7 +16,7 @@ pub extern "C" fn connect(socket: c_int, address: *const sockaddr, len: socklen_
         {
             #[cfg(target_os = "linux")]
             if open_coroutine_iouring::version::support_io_uring() {
-                return open_coroutine_core::event_loop::EventLoops::connect(
+                return open_coroutine_core::net::event_loop::EventLoops::connect(
                     Some(Lazy::force(&CONNECT)),
                     socket,
                     address,
