+++ open-coroutine/src/lib.rs
@@ -46,7 +46,7 @@
     clippy::single_char_lifetime_names, // TODO: change lifetime names
 )]
 
-pub use open_coroutine_core::config::Config;
+pub use open_coroutine_core::net::config::Config;
 pub use open_coroutine_macros::*;
 
 pub mod join;
